requirements.txt
# IIT Student Assistant — Python dependencies
# Install: pip install -r requirements.txt
# Tested with Python 3.11+

# Elasticsearch client
elasticsearch>=8.12.0,<9.0.0

# Embedding + reranking models
sentence-transformers>=5.0.0
torch>=2.0.0               # required by sentence-transformers

# Routing (cosine similarity over prototype embeddings)
scikit-learn>=1.0.0
numpy>=1.24.0

# HTTP (LLM SSE client)
requests>=2.28.0

# Environment / config
python-dotenv>=1.0.0

# Web UI (optional — only needed for app/streamlit_app.py)
streamlit>=1.28.0
